adjusted the variables to happy thoughts and obeyed the criteria for Thought body
